requirements.txt
Django == 5.1.2
ANTLR == 4.13.2